feat(invoice): show the fabric's colour and type, and allow decimal quantities
Two things reported from the shop floor.

The fabric list printed only the name and price, so "AL MAS FABRIC AM01" and
"AM02" — different colours at the same price — were impossible to tell apart
on screen. The code is already part of the name; the list now adds the colour
and type, e.g. "AL MAS FABRIC AM01 — blue · cotton (KES 400)". The review step
shows the same detail.

Fabric is sold by the metre, but the quantity columns were whole numbers, so
2.5 metres had to be rounded. Line quantity and ridhaa quantity now carry two
decimal places, and the totals follow. Shelf items are deliberately still
whole units — their stock is counted in units and half a cap does not exist —
so a fractional quantity there is refused with a plain explanation.

Quantities print tidily: 2.5 rather than 2.50, on both the WhatsApp receipt and
the PDF.

Verified: 2.5 x (1000 + 400) = 3,500; 0.75 x 400 = 300; a shelf item at 1.5 is
refused.

+++ resources/js/composables/useLineTotal.js
@@ -8,14 +8,14 @@
  * quantity: it is a separate item written onto the line at invoice time.
  */
 export function ridhaaTotal(item) {
-    const qty = parseInt(item?.ridhaa_qty) || 0;
+    const qty = parseFloat(item?.ridhaa_qty) || 0;
     const price = parseFloat(item?.ridhaa_price) || 0;
     return qty * price;
 }
 
 export function lineItemTotal(item) {
     if (!item) return 0;
-    const qty = parseInt(item.quantity) || 1;
+    const qty = parseFloat(item.quantity) || 1;
 
     if (item.item_type === 'collection') {
         return (parseFloat(item.unit_price) || 0) * qty;
